feat: full implement strip in api and ecommerce_app still gather that with gettin payment type from api  and display to user to chose whitch one is preffere cach or strip or any one else

+++ api/Presentation/dto/OrderDto.cs
@@ -7,10 +7,10 @@ namespace api.Presentation.dto
         public Guid Id { get; set; }
         public decimal Longitude { get; set; }
         public decimal Latitude { get; set; }
-        public decimal TotalPrice { get; set; }
+        public long TotalPrice { get; set; }
         
         public string Symbol { get; set; }
-        public decimal DeliveryFee { get; set; }
+        public int DeliveryFee { get; set; }
         public String Name { get; set; }
         public String UserPhone { get; set; }
         public string Status { get; set; }
@@ -27,10 +27,10 @@ namespace api.Presentation.dto
         public Guid Id { get; set; }
         public decimal Longitude { get; set; }
         public decimal Latitude { get; set; }
-        public decimal TotalPrice { get; set; }
+        public long TotalPrice { get; set; }
         
         public string Symbol { get; set; }
-        public decimal DeliveryFee { get; set; }
+        public  int DeliveryFee { get; set; }
         public String Name { get; set; }
         public String UserPhone { get; set; }
         public int Status { get; set; }
@@ -41,7 +41,7 @@ namespace api.Presentation.dto
     {
         [Required] public decimal Longitude { get; set; }
         [Required] public decimal Latitude { get; set; }
-        [Required] public decimal TotalPrice { get; set; }
+        [Required] public short TotalPrice { get; set; }
         public required string Symbol { get; set; }
 
         [Required] public List<CreateOrderItemDto> Items { get; set; }
